Waive signing entirely when the toggle is off
- Require signed commits before pushing only skipped the push
  check; rebases still passed `--gpg-sign` and a stack whose tips
  read unsigned still demanded a rebase. Off now means off:
  rebases drop the flag (never passing `--no-gpg-sign`, so plain
  git behaviour stands), nothing checks signatures and the rebase
  button stops promising signing it will not do
- The rebase target still prefers the branch's own origin ref, to
  keep pushed history's hashes; verifying the pushed commits only
  gates that choice while signing is required
- `rebaseNeed` and the target move to `SessionService+Rebase.swift`
  for file length
